Inline textMatches() helper function
