+++ lib/assertions.js
@@ -1,7 +1,6 @@
 import exists from './assertions/exists';
 import focused from './assertions/focused';
 import notFocused from './assertions/not-focused';
-import textMatches from './assertions/text-matches';
 
 const elementToString = require('./helpers/element-to-string');
 
@@ -104,7 +103,18 @@ export default class DOMAssertions {
    * assert.dom('.foo').textMatches(/[12]\d{3}/);
    */
   textMatches(regex, message) {
-    textMatches.call(this, regex, message);
+    let element = this.findTargetElement();
+    if (!element) return;
+
+    let result = regex.test(element.textContent);
+    let actual = element.textContent;
+    let expected = regex;
+
+    if (!message) {
+      message = `Element ${this.targetDescription} matches ${regex}`;
+    }
+
+    this.pushResult({ result, actual, expected, message });
   }
 
   /**
